Moved tests to right package
